+++ src/components/product/ProductWishButton.tsx
@@ -1,8 +1,7 @@
 import styled from '@emotion/styled';
 import { useProductWish } from '@/hooks/useProduct';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { FaHeart, FaRegHeart } from 'react-icons/fa';
-import type { ProductWish } from '@/types/product';
+import { useProductWishMutation } from '@/hooks/useProductWishMutation';
 
 interface ProductWishButtonProps {
   productId: number;
@@ -23,40 +22,8 @@ const Count = styled.span`
 `;
 
 const ProductWishButton = ({ productId }: ProductWishButtonProps) => {
-  const queryClient = useQueryClient();
   const { data } = useProductWish(productId);
-
-  const mutation = useMutation({
-    mutationFn: () => Promise.resolve(),
-    onMutate: async () => {
-      await queryClient.cancelQueries({ queryKey: ['productWish', productId] });
-
-      const prevData = queryClient.getQueryData(['productWish', productId]);
-
-      queryClient.setQueryData(['productWish', productId], (old: ProductWish | undefined) => {
-        if (!old) {
-          return old;
-        }
-
-        const isWished = !old.isWished;
-        const wishCount = old.wishCount + (isWished ? 1 : -1);
-
-        return { ...old, isWished, wishCount };
-      });
-
-      return { prevData };
-    },
-    onError: (_err, _variables, context) => {
-      if (context?.prevData) {
-        queryClient.setQueryData(['productWish', productId], context.prevData);
-      }
-    },
-    /* React Query가 상태를 invalidate하면서 서버값을 다시 가져오기때문에 주석처리해두었습니다.
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ['productWish', productId] });
-    }
-    */
-  });
+  const mutation = useProductWishMutation(productId);
 
   if (!data) {
     return null;
